+++ src/scripts/background.js
@@ -1,5 +1,5 @@
 import bugsnagClient from './lib/bugsnag';
-import { escapeHtml, report, secToHHMM } from './lib/utils';
+import { escapeHtml, isTogglURL, report, secToHHMM } from './lib/utils';
 
 import Db from './lib/db';
 import Ga from './lib/ga';
@@ -55,34 +55,60 @@ window.TogglButton = {
     '<div id="toggl-button-edit-form">' +
     '<form autocomplete="off">' +
     '<a class="toggl-button toggl-button-edit-form-button {service} active" href="javascript:void(0)">Stop timer</a>' +
-    '<a id="toggl-button-hide">&times;</a>' +
-    '<div class="toggl-button-row" id="toggl-button-duration-row">' +
-    '<input name="toggl-button-duration" tabindex="100" type="text" pattern="[\\d]{2}:[\\d]{2}:[\\d]{2}" title="Please write the duration in the \'hh:mm:ss\' format." id="toggl-button-duration" class="toggl-button-input" value="" placeholder="00:00" autocomplete="off">' +
-    '</div>' +
-    '<div class="toggl-button-row">' +
-    '<input name="toggl-button-description" tabindex="101" type="text" id="toggl-button-description" class="toggl-button-input" value="" placeholder="(no description)" autocomplete="off">' +
-    '</div>' +
-    '<div class="toggl-button-row">' +
-    '<input name="toggl-button-project-filter" tabindex="102" type="text" id="toggl-button-project-filter" class="toggl-button-input" value="" placeholder="Filter Projects" autocomplete="off">' +
-    '<a href="javascript:void(0)" class="filter-clear">&times;</a>' +
-    '<div id="toggl-button-project-placeholder" class="toggl-button-input" disabled><span class="tb-project-bullet"></span><div class="toggl-button-text">Add project</div><span>▼</span></div>' +
-    '<div id="project-autocomplete">{projects}</div>' +
-    '</div>' +
-    '<div class="toggl-button-row">' +
-    '<input name="toggl-button-tag-filter" tabindex="103" type="text" id="toggl-button-tag-filter" class="toggl-button-input" value="" placeholder="Filter Tags" autocomplete="off">' +
-    '<a href="javascript:void(0)" class="add-new-tag">+ Add</a>' +
-    '<a href="javascript:void(0)" class="filter-clear">&times;</a>' +
-    '<div id="toggl-button-tag-placeholder" class="toggl-button-input" disabled><div class="toggl-button-text">Add tags</div><span>▼</span></div>' +
-    '<div id="tag-autocomplete">' +
-    '<div class="tag-clear">Clear selected tags</div>' +
-    '{tags}</div>' +
-    '</div>' +
-    '<div class="toggl-button-row tb-billable {billable}" tabindex="103">' +
+    '<a id="toggl-button-hide"></a>' +
+
+    `<div class="Dialog__field" id="toggl-button-duration-row">
+      <div>
+        <input name="toggl-button-duration" type="text" id="toggl-button-duration" class="Input" value="" placeholder="00:00" autocomplete="off">
+      </div>
+    </div>` +
+
+    `<div class="Dialog__field">
+      <div><input name="toggl-button-description" type="text" id="toggl-button-description" class="Input" value="" placeholder="What are you doing?" autocomplete="off" /></div>
+    </div>` +
+
+    `
+    <div class="Dialog__field" tabindex="0">
+      <div>
+        <div id="toggl-button-project-placeholder" class="FormFieldTrigger__trigger" disabled><span class="tb-project-bullet"><div>No project</div></span><span class="Popdown__caret"></span></div>
+        <div class="Popdown__overlay"></div>
+        <div class="Popdown__content">
+          <div class="Popdown__filterContainer">
+            <input name="toggl-button-project-filter" type="text" id="toggl-button-project-filter" class="Popdown__filter" value="" placeholder="Find project..." autocomplete="off">
+          </div>
+          <div id="project-autocomplete">{projects}</div>
+        </div>
+      </div>
+    </div>
+    ` +
+
+    `
+    <div class="Dialog__field" tabindex="0">
+      <div>
+        <div id="toggl-button-tag-placeholder" class="FormFieldTrigger__trigger" disabled><div>Add tags</div><span class="Popdown__caret"></span></div>
+        <div class="Popdown__overlay"></div>
+        <div class="Popdown__content">
+          <div class="Popdown__filterContainer">
+            <input name="toggl-button-tag-filter" type="text" id="toggl-button-tag-filter" class="Popdown__filter" value="" placeholder="Find tags..." autocomplete="off">
+          </div>
+          <div id="tag-autocomplete">
+            <div class="tag-clear">Clear selected tags</div>
+            <a href="javascript:void(0)" class="add-new-tag">+ Add tag "<span></span>"</a>
+            {tags}
+          </div>
+        </div>
+      </div>
+    </div>
+    ` +
+
+    '<div class="Dialog__field">' +
+    '<div class="tb-billable {billable}">' +
     '<div class="toggl-button-billable-label">Billable</div>' +
     '<div class="toggl-button-billable-flag"><span></span></div>' +
     '</div>' +
-    '<div id="toggl-button-update" tabindex="105">DONE</div>' +
-    '<input type="submit" class="hidden">' +
+    '</div>' +
+    '<div id="toggl-button-update" tabindex="0" class="Button__button">Done</div>' +
+    '<input type="submit" class="toggl-button-hidden">' +
     '</from>' +
     '</div>',
 
@@ -149,14 +175,15 @@ window.TogglButton = {
               localStorage.setItem('projects', JSON.stringify(projectMap));
               localStorage.setItem('clients', JSON.stringify(clientMap));
               TogglButton.$user = resp.data;
-              TogglButton.$user.time_entries = TogglButton.$user.time_entries.map((te) => {
-                // Ensure empty values from v8 become null.
-                return {
-                  ...te,
-                  pid: te.pid || null,
-                  tid: te.tid || null
-                };
-              });
+              TogglButton.$user.time_entries = (TogglButton.$user.time_entries || [])
+                .map((te) => {
+                  // Ensure empty values from v8 become null.
+                  return {
+                    ...te,
+                    pid: te.pid || null,
+                    tid: te.tid || null
+                  };
+                });
               TogglButton.$user.projectMap = projectMap;
               TogglButton.$user.clientMap = clientMap;
               TogglButton.$user.clientNameMap = clientNameMap;
@@ -541,6 +568,7 @@ window.TogglButton = {
               TogglButton.localEntry = entry;
               TogglButton.updateTriggers(entry);
               ga.reportEvent(timeEntry.type, timeEntry.service);
+              db.bumpTrackedCount();
             } else {
               error = xhr.responseText;
             }
@@ -1095,7 +1123,7 @@ window.TogglButton = {
 
   setupBillable: function () {
     if (TogglButton.canSeeBillable) {
-      return '" tabindex="103';
+      return '" ';
     }
 
     return 'no-billable" tabindex="-1';
@@ -1103,7 +1131,7 @@ window.TogglButton = {
 
   fillProjects: function () {
     let html =
-        '<p class="project-row" data-pid="0"><span class="tb-project-bullet tb-project-color tb-no-color"></span>No project</p>';
+        '<p class="project-row" data-pid="0"><span class="tb-project-bullet tb-project-color tb-no-color">No project</span></p>';
     const projects = TogglButton.$user.projectMap;
     const clients = TogglButton.$user.clientMap;
     const clientNames = TogglButton.$user.clientNameMap;
@@ -1137,9 +1165,9 @@ window.TogglButton = {
           '"><li class="ws-row' +
           hideWs +
           '" title="' +
-          escapeHtml(element.name.toUpperCase()) +
+          escapeHtml(element.name) +
           '">' +
-          escapeHtml(element.name.toUpperCase()) +
+          escapeHtml(element.name) +
           '</li>' +
           '<ul class="client-list" data-cid="0">';
       });
@@ -1204,16 +1232,16 @@ window.TogglButton = {
         escapeHtml(project.name) +
         '" data-pid="' +
         project.id +
-        '"><span class="tb-project-bullet tb-project-color" style="background-color: ' +
+        '"><span class="tb-project-bullet tb-project-color" style="color: ' +
         project.hex_color +
-        '"></span>' +
+        '">' +
         '<span class="item-name ' +
         hasTasks +
         '" title="' +
         escapeHtml(project.name) +
         '">' +
         escapeHtml(project.name) +
-        '</span>';
+        '</span></span>';
 
     if (tasks) {
       tasksCount = tasks.length + ' task';
@@ -1861,7 +1889,10 @@ window.TogglButton = {
         } else if (request.type === 'logout') {
           TogglButton.logoutUser().then(resolve);
         } else if (request.type === 'sync') {
-          TogglButton.fetchUser();
+          const res = TogglButton.fetchUser();
+          if (request.respond) {
+            resolve(res);
+          }
         } else if (request.type === 'timeEntry') {
           TogglButton.createTimeEntry(request)
             .then((response) => {
@@ -1936,7 +1967,7 @@ window.TogglButton = {
   },
 
   tabUpdated: async function (tabId, changeInfo, tab) {
-    if (!TogglButton.$user) {
+    if (!TogglButton.$user && !isTogglURL(tab.url)) {
       TogglButton.setBrowserActionBadge();
       return;
     }
@@ -1944,7 +1975,7 @@ window.TogglButton = {
       changeInfo.status === 'complete' &&
       tab.url.indexOf('chrome://') === -1
     ) {
-      const domain = await TogglButton.extractDomain(tab.url);
+      const domain = await TogglButton.extractDomain(tab.url, false);
       const permission = { origins: domain.origins };
 
       if (process.env.DEBUG) {
@@ -1987,7 +2018,10 @@ window.TogglButton = {
     }
     browser.tabs.insertCSS(tabId, { file: 'styles/style.css' })
       .then(function () {
-        browser.tabs.insertCSS(tabId, { file: 'styles/autocomplete.css' });
+        return browser.tabs.insertCSS(tabId, { file: 'styles/edit-form.css' });
+      })
+      .then(function () {
+        return browser.tabs.insertCSS(tabId, { file: 'styles/autocomplete.css' });
       });
 
     browser.tabs.executeScript(tabId, { file: 'scripts/common.js' })
@@ -1998,9 +2032,9 @@ window.TogglButton = {
       });
   },
 
-  extractDomain: async function (url) {
+  extractDomain: async function (url, checkLogin = true) {
     let domain;
-    if (!TogglButton.$user) {
+    if (checkLogin && !TogglButton.$user) {
       return false;
     }
     // find & remove protocol (http, ftp, etc.) and get domain
@@ -2083,7 +2117,9 @@ window.TogglButton = {
           show = show != null ? show : 2;
           db.set('settings-active-tab', 2);
           db.set('show-permissions-info', show);
-          browser.runtime.openOptionsPage();
+          if (TogglButton.$user) {
+            browser.runtime.openOptionsPage();
+          }
         }
       });
     }
@@ -2172,6 +2208,10 @@ browser.windows.onCreated.addListener(function () {
 });
 browser.windows.onRemoved.addListener(TogglButton.stopTrackingOnBrowserClosed);
 
+browser.runtime
+  .setUninstallURL('https://toggl.com/toggl-button-feedback/')
+  .catch(bugsnagClient.notify);
+
 window.onbeforeunload = function () {
   db.get('stopAutomatically')
     .then((stopAutomatically) => {
@@ -2188,7 +2228,11 @@ if (!FF) {
 // Check whether new version is installed
 browser.runtime.onInstalled.addListener(function (details) {
   if (details.reason === 'install') {
-    if (!FF) TogglButton.checkPermissions(0);
+    if (!TogglButton.$user) {
+      browser.tabs.create({ url: 'html/login.html?source=install' });
+    } else if (!FF) {
+      TogglButton.checkPermissions(0);
+    }
   } else if (details.reason === 'update') {
     console.info(`Updated from ${details.previousVersion} to ${process.env.VERSION}.`);
     const [ prevMajor, prevMinor ] = details.previousVersion.split('.').map(Number);