+++ src/scripts/content/atlassian.js
@@ -5,6 +5,10 @@ togglbutton.render(
   '#ghx-detail-view [spacing] h1:not(.toggl)',
   { observe: true },
   function () {
+    if (process.env.DEBUG) {
+      console.info('🏃 "Jira 2017 sidebar" rendering');
+    }
+
     const rootElem = $('#ghx-detail-view');
     const container = createTag('div', 'jira-ghx-toggl-button');
     const titleElem = $('[spacing] h1', rootElem);
@@ -27,65 +31,46 @@ togglbutton.render(
   }
 );
 
-// Jira 2018-X Sprint Modal
-// We select dialog content in order to wait for the SPA to render.
-// N.B. this can bring its own issues (see comment about infinite re-renders).
-// The h1 element gets replaced and no longer has .toggl class, so we have to be careful.
+// Jira 2018-11 issue page and board page single issue modal. Uses functions for timer values due to SPA on issue-lists.
 togglbutton.render(
-  'div[role="dialog"] h1:not(.toggl)',
-  { observe: true },
-  function (needle) {
-    const root = needle.closest('div[role="dialog"]');
-    const id = $('div:last-child > a[spacing="none"][href^="/browse/"]:last-child', root);
-    const description = $('h1:first-child', root);
-    let project = $('[data-test-id="navigation-apps.project-switcher-v2"] button > div:nth-child(2) > div');
-    let link;
+  // The main "issue link" at the top of the issue.
+  // Extra target and role selectors are to avoid picking up wrong links on issue-list-pages.
+  'a[href^="/browse/"][target=_blank]:not([role=list-item]):not(.toggl)',
 
-    if (project === null) {
-      project = $('a[href^="/browse/"][target=_self]');
-    }
-
-    if (id !== null && description !== null) {
-      link = togglbutton.createTimerLink({
-        className: 'jira2018',
-        description: id.textContent + ' ' + description.textContent,
-        projectName: project && project.textContent
-      });
-
-      // Link is not placed in exactly the same element as a regular issue page,
-      // else we encounter infinite re-renders when the SPA updates the DOM.
-      id.parentNode.appendChild(link);
-    }
-  }
-);
-
-// Jira 2018-11 issue page. Uses functions for timer values due to SPA on issue-lists.
-togglbutton.render(
-  '#jira-frontend:not(.toggl)',
   { observe: true },
   function (elem) {
     let titleElement;
     let projectElement;
 
-    // The main "issue link" at the top of the issue.
-    // Extra target and role selectors are to avoid picking up wrong links on issue-list-pages.
-    const issueNumberElement = $('a[href^="/browse/"][target=_blank]:not([role=list-item])', elem);
+    const issueNumberElement = elem;
     const container = issueNumberElement.parentElement.parentElement.parentElement;
+
+    if (container.querySelector('.toggl-button')) {
+      // We're checking for existence of the button as re-rendering in Jira SPA is not reliable for our uses.
+      if (process.env.DEBUG) {
+        console.info('🚫 "Jira 2018-11 issue page and board page" quit rendering early');
+      }
+      return;
+    }
+
+    if (process.env.DEBUG) {
+      console.info('🏃 "Jira 2018-11 issue page and board page" rendering');
+    }
 
     function getDescription () {
       let description = '';
 
       // Title/summary of the issue - we use the hidden "edit" button that's there for a11y
       // in order to avoid picking up actual page title in the case of issue-list-pages.
-      titleElement = $('h1 ~ button[aria-label]', elem).previousSibling;
+      titleElement = document.querySelector('h1 ~ button[aria-label]');
 
       if (issueNumberElement) {
         description += issueNumberElement.textContent.trim();
       }
 
-      if (titleElement) {
+      if (titleElement && titleElement.previousSibling) {
         if (description) description += ' ';
-        description += titleElement.textContent.trim();
+        description += titleElement.previousSibling.textContent.trim();
       }
 
       return description;
@@ -122,6 +107,10 @@ togglbutton.render(
   '.issue-header-content:not(.toggl)',
   { observe: true },
   function (elem) {
+    if (process.env.DEBUG) {
+      console.info('🏃 "Jira 2017 issue page" rendering');
+    }
+
     const numElem = $('#key-val', elem);
     const titleElem = $('#summary-val', elem) || '';
     let projectElem = $('.bgdPDV');