+++ src/scripts/content/toggl.js
@@ -7,7 +7,7 @@ const offlineUser = localStorage.getItem('offline_users');
 if (offlineUser) {
   userData = JSON.parse(localStorage.getItem('offline_users-' + offlineUser));
   if (userData && userData.offlineData) {
-    browser.extension.sendMessage({
+    browser.runtime.sendMessage({
       type: 'userToken',
       apiToken: userData.offlineData.api_token
     });
@@ -24,8 +24,27 @@ if (offlineUser) {
 
 document.addEventListener('webkitvisibilitychange', function () {
   if (!document.webkitHidden) {
-    browser.extension.sendMessage({ type: 'sync' });
+    browser.runtime.sendMessage({ type: 'sync' });
   }
 });
 
-browser.extension.sendMessage({ type: 'sync' });
+browser.runtime.sendMessage({ type: 'sync' });
+
+function completeLogin () {
+  location.href = browser.runtime.getURL('html/login.html?source=web-login');
+}
+
+function handleIncomingMessage (event) {
+  if (
+    event.source === window &&
+    event.data &&
+    event.data.direction === 'from-public-web'
+  ) {
+    switch (event.data.message) {
+      case 'login-success': completeLogin(); break;
+      default: console.log('Unsupported event', event);
+    }
+  }
+}
+
+window.addEventListener('message', handleIncomingMessage);