+++ src/scripts/origins.js
@@ -22,12 +22,14 @@ export default {
   },
   'atlassian.com': {
     url: '*://*.atlassian.com/*',
-    name: 'Atlassian'
+    name: 'Atlassian / Jira',
+    file: 'atlassian.js'
   },
   'atlassian.net': {
     url: '*://*.atlassian.net/*',
-    name: 'Atlassian',
-    clone: 'true'
+    name: 'Atlassian / Jira',
+    clone: 'true',
+    file: 'atlassian.js'
   },
   'attask-ondemand.com': {
     url: '*://*.attask-ondemand.com/*',
@@ -157,6 +159,10 @@ export default {
   'feedly.com': {
     url: '*://*.feedly.com/*',
     name: 'Feedly'
+  },
+  'fibery.io': {
+    url: '*://*.fibery.io/*',
+    name: 'Fibery'
   },
   'focuster.com': {
     url: '*://*.focuster.com/*',
@@ -362,8 +368,8 @@ export default {
     name: 'ProcessWire'
   },
   'produck.io': {
-    'url': '*://*.produck.io/*',
-    'name': 'Produck'
+    url: '*://*.produck.io/*',
+    name: 'Produck'
   },
   'producteev.com': {
     url: '*://*.producteev.com/*',