+++ src/scripts/popup.js
@@ -2,10 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { isSameDay } from 'date-fns';
 
+import { secToHhmmImproved } from './@toggl/time-format-utils';
 import Summary from './components/Summary';
 import TimeEntriesList from './components/TimeEntriesList';
 import Timer, { formatDuration } from './components/Timer';
 import { ProjectAutoComplete, TagAutoComplete } from './lib/autocomplete';
+import { parseDuration } from './lib/timerUtils';
+import renderLogin from './initializers/login';
 
 const browser = require('webextension-polyfill');
 
@@ -39,7 +42,7 @@ window.PopUp = {
   $header: document.querySelector('.header'),
   $menuView: document.querySelector('#menu'),
   $editView: document.querySelector('#entry-form'),
-  $loginView: document.querySelector('#login-form'),
+  $loginView: document.querySelector('#login-view'),
   $revokedWorkspaceView: document.querySelector('#revoked-workspace'),
   $entries: document.querySelector('.entries-list'),
   defaultErrorMessage: 'Error connecting to server',
@@ -66,8 +69,6 @@ window.PopUp = {
               JSON.stringify(TogglButton.$latestStoppedEntry)
             );
           }
-        } else {
-          PopUp.showCurrentDuration(true);
         }
         if (!PopUp.$header.getAttribute('data-view')) {
           PopUp.switchView(PopUp.$menuView);
@@ -215,23 +216,14 @@ window.PopUp = {
   },
 
   switchView: function (view) {
+    if (view === PopUp.$loginView) {
+      renderLogin(PopUp.$loginView, true);
+    }
     PopUp.$header.setAttribute('data-view', view.id);
   },
 
   formatMe: function (n) {
     return n < 10 ? '0' + n : n;
-  },
-
-  msToTime: function (duration) {
-    let seconds = parseInt((duration / 1000) % 60, 10);
-    let minutes = parseInt((duration / (1000 * 60)) % 60, 10);
-    let hours = parseInt(duration / (1000 * 60 * 60), 10);
-
-    hours = hours < 10 ? '0' + hours : hours;
-    minutes = minutes < 10 ? '0' + minutes : minutes;
-    seconds = seconds < 10 ? '0' + seconds : seconds;
-
-    return hours + ':' + minutes + ':' + seconds;
   },
 
   /* Edit form functions */
@@ -247,8 +239,9 @@ window.PopUp = {
     togglButtonDescription.value = TogglButton.$curEntry.description
       ? TogglButton.$curEntry.description
       : '';
-    togglButtonDuration.value = PopUp.msToTime(
-      new Date() - new Date(TogglButton.$curEntry.start)
+    togglButtonDuration.value = secToHhmmImproved(
+      new Date() - new Date(TogglButton.$curEntry.start),
+      { html: false }
     );
 
     PopUp.$projectAutocomplete.setup(pid, tid);
@@ -263,11 +256,12 @@ window.PopUp = {
     togglButtonDescription.scrollLeft = 0;
 
     PopUp.durationChanged = false;
-    PopUp.showCurrentDuration(true);
+    PopUp.updateDurationInput(true);
   },
 
-  showCurrentDuration: function (startTimer) {
+  updateDurationInput: function (startTimer) {
     if (TogglButton.$curEntry === null) {
+      PopUp.stopDurationInput();
       return;
     }
 
@@ -275,22 +269,21 @@ window.PopUp = {
     const durationField = document.querySelector('#toggl-button-duration');
 
     // Update edit form duration field
-    if (
-      durationField !== document.activeElement &&
-      PopUp.durationChanged === false
-    ) {
+    if (PopUp.durationChanged === false) {
       durationField.value = duration;
-    } else {
-      PopUp.durationChanged = true;
     }
 
     if (startTimer) {
-      if (!PopUp.$timer) {
-        PopUp.$timer = setInterval(function () {
-          PopUp.showCurrentDuration();
-        }, 1000);
-      }
+      PopUp.stopDurationInput();
+      PopUp.$timer = setInterval(function () {
+        if (process.env.DEBUG) console.log('🕒🐭 Tick tock, the mouse ran up the clock..');
+        PopUp.updateDurationInput();
+      }, 1000);
     }
+  },
+
+  stopDurationInput: function () {
+    clearInterval(PopUp.$timer);
   },
 
   updateBillable: function (pid, noOverwrite) {
@@ -324,26 +317,12 @@ window.PopUp = {
     }
   },
 
-  // Duration changed let's calculate new start
-  getStart: function () {
-    const arr = document.querySelector('#toggl-button-duration').value.split(':');
-
-    if (!PopUp.isNumber(arr.join(''))) {
-      return false;
-    }
-
-    const duration = 1000 * (arr[2] || 0) + 60000 * arr[1] + 3600000 * arr[0];
-
-    const now = new Date();
-    return new Date(now.getTime() - duration);
-  },
-
   isNumber: function (n) {
     return !isNaN(parseFloat(n)) && isFinite(n);
   },
 
   toggleBillable: function (visible) {
-    const tabIndex = visible ? '103' : '-1';
+    const tabIndex = visible ? '0' : '-1';
     PopUp.$billable.setAttribute('tabindex', tabIndex);
     PopUp.$billable.classList.toggle('no-billable', !visible);
   },
@@ -353,12 +332,22 @@ window.PopUp = {
     PopUp.$billable.classList.toggle('tb-checked', billable);
   },
 
-  submitForm: function (that) {
+  submitForm: function () {
+    PopUp.stopDurationInput();
+
+    // Translate human duration input if submitted without blurring
+    const $duration = document.querySelector('#toggl-button-duration');
+    let duration = $duration.value;
+    if (duration) {
+      duration = parseDuration(duration).asSeconds();
+      $duration.value = secToHhmmImproved(duration, { html: false });
+    }
+
     if (!this.isformValid()) {
       return;
     }
-    const selected = PopUp.$projectAutocomplete.getSelected();
 
+    const selected = PopUp.$projectAutocomplete.getSelected();
     const billable = !!document.querySelector(
       '.tb-billable.tb-checked:not(.no-billable)'
     );
@@ -375,15 +364,13 @@ window.PopUp = {
       service: 'dropdown'
     };
 
-    const start = PopUp.getStart();
-
-    if (start) {
+    if (duration) {
+      const start = new Date((new Date()).getTime() - duration * 1000);
       request.start = start.toISOString();
       request.duration = -1 * Math.floor(start.getTime() / 1000);
     }
 
     PopUp.sendMessage(request);
-    PopUp.renderTimer();
     PopUp.switchView(PopUp.$menuView);
   },
 
@@ -420,6 +407,19 @@ window.PopUp = {
       .addEventListener('submit', function (e) {
         PopUp.submitForm(this);
         e.preventDefault();
+      });
+
+    document
+      .querySelector('#toggl-button-project-placeholder')
+      .closest('.Dialog__field')
+      .addEventListener('focus', (e) => {
+        PopUp.$projectAutocomplete.openDropdown();
+      });
+    document
+      .querySelector('#toggl-button-tag-placeholder')
+      .closest('.Dialog__field')
+      .addEventListener('focus', (e) => {
+        PopUp.$tagAutocomplete.openDropdown();
       });
 
     PopUp.$projectAutocomplete.onChange(function (selected) {
@@ -505,20 +505,6 @@ document.addEventListener('DOMContentLoaded', function () {
       });
 
     document
-      .querySelector('#signin')
-      .addEventListener('submit', function (event) {
-        event.preventDefault();
-        PopUp.$errorLabel.classList.remove('show');
-        const request = {
-          type: 'login',
-          respond: true,
-          username: document.querySelector('#login_email').value,
-          password: document.querySelector('#login_password').value
-        };
-        PopUp.sendMessage(request);
-      });
-
-    document
       .querySelector('#workspace')
       .addEventListener('submit', function (event) {
         event.preventDefault();
@@ -532,6 +518,23 @@ document.addEventListener('DOMContentLoaded', function () {
           workspace
         };
         PopUp.sendMessage(request);
+      });
+
+    document
+      .querySelector('#toggl-button-duration')
+      .addEventListener('keydown', function (event) {
+        // Doesn't cover all cases; can't really do it without introducing more state.
+        // Need a refactor.
+        if (event.code !== 'Enter' && event.code !== 'Tab') {
+          PopUp.durationChanged = true;
+        }
+      });
+    document
+      .querySelector('#toggl-button-duration')
+      .addEventListener('blur', function (event) {
+        const value = event.target.value || '';
+        const parsedInput = parseDuration(value).asSeconds();
+        event.target.value = secToHhmmImproved(parsedInput, { html: false });
       });
 
     PopUp.$entries.addEventListener('click', function (e) {
@@ -549,6 +552,7 @@ document.addEventListener('DOMContentLoaded', function () {
       };
 
       PopUp.sendMessage(request);
+      window.scrollTo(0, 0);
     });
   } catch (e) {
     browser.runtime.sendMessage({